pyproject.toml
[build-system]
requires = ["setuptools >= 61.0", "setuptools-git-versioning"]
build-backend = "setuptools.build_meta"

[project]
dynamic = ["version"]
name = "yadg"
authors = [
    {name = "Peter Kraus", email = "peter.kraus@tu-berlin.de"},
    {name = "Nicolas Vetsch"},
]
maintainers = [
    {name = "Peter Kraus", email = "peter.kraus@tu-berlin.de"},
]
description = "yet another datagram"
readme = "README.md"
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Operating System :: OS Independent",
]
requires-python = ">= 3.9"
dependencies = [
    "numpy >= 1.20",
    "pint >= 0.22",
    "pyyaml",
    "uncertainties",
    "striprtf",
    "tzlocal",
    "packaging",
    "python-dateutil",
    "openpyxl >= 3.0.0",
    "olefile >= 0.47",
    "h5netcdf ~= 1.0",
    "pandas >= 2.0",
    "babel >= 2.15",
    "xarray-datatree >= 0.0.12",
    "dgbowl-schemas @ git+https://github.com/dgbowl/dgbowl-schemas.git@dataschema_5.1",
]

[project.optional-dependencies]
testing = ["pytest"]
docs = [
    "sphinx ~= 7.2",
    "sphinx-rtd-theme ~= 1.3.0",
    "autodoc-pydantic ~= 2.1.0",
]

[project.urls]
Documentation = "https://dgbowl.github.io/yadg/master"
Issues = "https://github.com/dgbowl/yadg/issues"
Repository = "https://github.com/dgbowl/yadg/"

[project.scripts]
yadg = "yadg:run_with_arguments"

[tool.setuptools-git-versioning]
enabled = true
dev_template = "{tag}.dev{ccount}"
dirty_template = "{tag}.dev{ccount}"

[tool.ruff]